# internal/models/generated.go
// Package models provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.5.0 DO NOT EDIT.
package models

import (
	"time"
)

// Error defines model for Error.
type Error struct {
	// Error Error message
	Error *string `json:"error,omitempty"`
}

// Resource defines model for Resource.
type Resource struct {
	// Active Whether the resource is active
	Active bool `json:"active"`

	// CreatedAt When the resource was created
	CreatedAt time.Time `json:"created_at"`

	// DeletedAt When the resource was deleted (soft delete)
	DeletedAt *time.Time `json:"deleted_at"`

	// Description Description of the resource
	Description *string `json:"description,omitempty"`

	// Id Unique identifier for the resource
	Id int64 `json:"id"`

	// Name Name of the resource
	Name string `json:"name"`

	// UpdatedAt When the resource was last updated
	UpdatedAt time.Time `json:"updated_at"`
}

// ResourceCreateRequest defines model for ResourceCreateRequest.
type ResourceCreateRequest struct {
	// Description Description of the resource
	Description *string `json:"description,omitempty"`

	// Name Name of the resource
	Name string `json:"name"`
}

// ResourceUpdateRequest defines model for ResourceUpdateRequest.
type ResourceUpdateRequest struct {
	// Active Whether the resource is active
	Active *bool `json:"active,omitempty"`

	// Description Description of the resource
	Description *string `json:"description,omitempty"`

	// Name Name of the resource
	Name *string `json:"name,omitempty"`
}

// GetApiV1ResourcesParams defines parameters for GetApiV1Resources.
type GetApiV1ResourcesParams struct {
	// Active Filter by active status
	Active *bool `form:"active,omitempty" json:"active,omitempty"`
}

// PostApiV1ResourcesJSONRequestBody defines body for PostApiV1Resources for application/json ContentType.
type PostApiV1ResourcesJSONRequestBody = ResourceCreateRequest

// PutApiV1ResourcesIdJSONRequestBody defines body for PutApiV1ResourcesId for application/json ContentType.
type PutApiV1ResourcesIdJSONRequestBody = ResourceUpdateRequest